Jobbar på skappa order metod

+++ src/Planhus.java
@@ -1,6 +1,6 @@
 public class Planhus extends Fastighet {
 
-    Planhus(int planYta, int antalRum, int försäljningsPris, float tomtPris, int tomtYta){
+    Planhus(int planYta, int antalRum, float försäljningsPris, float tomtPris, int tomtYta){
 
         this.planYta = planYta;
         this.antalRum = antalRum;
